+++ CHANGELOG.md
@@ -12,11 +12,19 @@
 
 ### Misc
 
+- The `IOS16_4_First_Responder_Bug_CollectionView` workaround is longer capped to iOS versions less than 19.0. It was verified as required in the iOS 19 beta.
+
 ### Internal
 
 - Added the `ScrollCompletionHandlerViewController` demo to showcase the new completion handler with different layout configurations.
 
 # Past Releases
+
+# [16.1.0] - 2025-06-17
+
+### Changed
+
+- Blueprint dependency updated to 6.0.
 
 # [16.0.4] - 2025-06-06
 
@@ -1174,8 +1182,9 @@ listActions.scrolling.scrollToSection(
 Earlier releases were ad-hoc and not tracked. To see all changes, please reference [closed PRs on Github](https://github.com/kyleve/Listable/pulls?q=is%3Apr+is%3Aclosed).
 
 
-[Main]: https://github.com/square/Listable/compare/16.0.4...main
-[16.0.3]: https://github.com/square/Listable/compare/16.0.4...16.0.3
+[Main]: https://github.com/square/Listable/compare/16.1.0...main
+[16.1.0]: https://github.com/square/Listable/compare/16.1.0...16.0.4
+[16.0.4]: https://github.com/square/Listable/compare/16.0.4...16.0.3
 [16.0.3]: https://github.com/square/Listable/compare/16.0.3...16.0.2
 [16.0.2]: https://github.com/square/Listable/compare/16.0.2...16.0.1
 [16.0.1]: https://github.com/square/Listable/compare/16.0.1...16.0.0